server/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
python-dotenv==1.1.1
Werkzeug==3.1.3
gunicorn==21.2.0
